+++ parallax/__init__.py
@@ -1,5 +1,5 @@
 from . import spaces
-from .core import Env, State, VectorEnv
+from .core import Agents, Env, MARLEnv, MARLState, MARLVectorEnv, State, VectorEnv
 from .spaces import Space
 from .wrappers import AutoResetWrapper, TimeLimit, VmapWrapper, Wrapper
 
@@ -7,6 +7,10 @@ __all__ = [
     "Env",
     "VectorEnv",
     "State",
+    "MARLEnv",
+    "MARLVectorEnv",
+    "MARLState",
+    "Agents",
     "Space",
     "spaces",
     "AutoResetWrapper",